Rename Test Files to Reflect New Class ApiConnection
